requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
deluge-client==1.9.0
Django==3.2.8
django-ajax-datatable==4.3.1
django-background-tasks==1.2.5
django-compat==1.0.15
django-widget-tweaks==1.4.9
idna==3.3
parse-torrent-title==2.3
python-dateutil==2.8.2
pytz==2021.3
qbittorrent-api==2021.8.23
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
transmission-rpc==3.3.0
typing-extensions==3.10.0.2
urllib3==1.26.7
